+++ apps/crates/service_hub/organization/src/service/rank.rs
@@ -68,7 +68,7 @@ impl RankService {
             sort: Set(req.sort),
             desc: Set(req.desc),
             level: Set(req.level),
-            status: Set(req.status as i8),
+            status: Set(req.status),
             ..Default::default()
         };
         let rank = self
@@ -97,7 +97,7 @@ impl RankService {
             level: Set(req.level),
             sort: Set(req.sort),
             desc: Set(req.desc),
-            status: Set(req.status as i8),
+            status: Set(req.status),
             ..Default::default()
         };
 
@@ -158,7 +158,7 @@ impl RankService {
     /// 更新数据状态
     pub async fn update_status(&self, req: UpdateRankStatusReq) -> Result<(), ErrorMsg> {
         self.rank_dao
-            .update_status(req.id, req.status as i8)
+            .update_status(req.id, req.status)
             .await
             .map_err(|err| {
                 error!("更新职级状态失败, err: {:#?}", err);
